E-courrier v4

+++ app/Http/Controllers/NatureController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Nature;
+use App\Models\Journal;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class NatureController extends Controller
 {
@@ -20,20 +22,78 @@ class NatureController extends Controller
         return back()->with('insert', 'nature ajouter avec success');
     }
 
-    public function edit(int $id)
+    public function edit(int $id, Nature $nature)
     {
+        $this->authorize('update', $nature);
         $nature = Nature::find($id);
-        return view('nature_update', compact(["nature"]));
+        return view('nature.update', compact(["nature"]));
     }
-    public function update(Request $request)
+
+    public function update(Request $request, Nature $nature)
     {
+        $this->authorize('update', $nature);
         $nature = nature::find($request->id);
         $nature->nom = $request->nom;
         $nature->save();
         return back()->with('update', 'nature mise à jour avec success');
     }
-    public function delete(int $id)
+
+
+     // corbeille dashboard
+     public function corbeille()
+     {
+         $rows = Nature::onlyTrashed()->get();
+         return view('nature.corbeille', compact(['rows']));
+     }
+
+     // restaurer tous un element
+     public function restore(int $id, Nature $nature)
+     {
+         $this->authorize('restore', $nature);
+         $delete = Nature::where('id', $id)->restore();
+         // check data restore or not
+         if ($delete == 1) {
+             $success = true;
+             $message = "Nature restaurer avec success";
+         } else {
+             $success = true;
+             $message = "Nature non trouver";
+         }
+
+         //  return response
+         return response()->json([
+             'success' => $success,
+             'message' => $message,
+         ]);
+     }
+
+     // restaurer tous les elements
+     public function restore_all()
+     {
+         $delete = Nature::withTrashed()->restore();
+         // check data restore or not
+         if ($delete == 1) {
+             $success = true;
+             $message = "Tout les nature ont été restaurés avec success";
+         } else {
+             $success = true;
+             $message = "La corbeille a été vider";
+         }
+
+         //  return response
+         return response()->json([
+             'success' => $success,
+             'message' => $message,
+         ]);
+     }
+
+    public function delete(int $id, Nature $nature)
     {
+        $this->authorize('delete', $nature);
+        $journal = new Journal();
+        $journal->user_id = Auth::user()->id;
+        $journal->libelle = 'Suppression de la nature N°' . $id;
+        $journal->save();
         // si oui supprimer de la BD
         $delete = Nature::destroy($id);
         // check data deleted or not
